refactor(benchmarks): isolate low-level targets
